+++ packages/storybook-addon-featureflags/src/constants.js
@@ -1,8 +1,0 @@
-export const ADDON_ID = 'ps-design-system-storybook-addon-featureflags'
-export const PANEL_ID = `${ADDON_ID}/panel`
-
-export const EVENTS = {
-  change: `${ADDON_ID}:change`
-}
-
-export const featFlags = ['psds2020Colors']