+++ packages/navuser/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.11](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.1.10...@pluralsight/ps-design-system-navuser@1.1.11) (2020-09-15)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
+## [1.1.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.1.9...@pluralsight/ps-design-system-navuser@1.1.10) (2020-09-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
+## [1.1.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.1.8...@pluralsight/ps-design-system-navuser@1.1.9) (2020-09-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
 ## [1.1.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.1.7...@pluralsight/ps-design-system-navuser@1.1.8) (2020-09-09)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-navuser