+++ packages/site/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [21.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@21.0.2...@pluralsight/ps-design-system-site@21.0.3) (2020-09-15)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+## [21.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@21.0.1...@pluralsight/ps-design-system-site@21.0.2) (2020-09-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+## [21.0.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@21.0.0...@pluralsight/ps-design-system-site@21.0.1) (2020-09-10)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
 # [21.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@20.0.11...@pluralsight/ps-design-system-site@21.0.0) (2020-09-09)