+++ packages/text/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.0.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-text@16.0.20...@pluralsight/ps-design-system-text@16.0.21) (2020-09-15)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-text
+
+
+
+
+
+## [16.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-text@16.0.19...@pluralsight/ps-design-system-text@16.0.20) (2020-09-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-text
+
+
+
+
+
 ## [16.0.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-text@16.0.18...@pluralsight/ps-design-system-text@16.0.19) (2020-09-09)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-text