+++ packages/tooltip/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.0.13...@pluralsight/ps-design-system-tooltip@6.0.14) (2020-09-15)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
+## [6.0.13](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.0.12...@pluralsight/ps-design-system-tooltip@6.0.13) (2020-09-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
 ## [6.0.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.0.11...@pluralsight/ps-design-system-tooltip@6.0.12) (2020-08-10)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-tooltip